+++ CHANGELOG.md
@@ -6,7 +6,9 @@ principles.
 - Added: config option `chief.preview-mode` to set the default admin preview behavior. 
 - Added: filefield now shows thumbnail of image if an image was uploaded.
 - Added: Pagebuilder module and text sections now have a online/offline toggle. This allows for the admin to toggle the frontend visibility of certain sections.
-This requires a migration to be run.
+- Added: sortable option for models. Add a $useManualSorting on a PageManager to allow the admin to sort their models. Note: for the time being it only works for non-paginated index views.
+
+This release requires migrations to be run. 
 
 ## 0.5.9 - 2020-10-26
 - Fixed: issue where visiting archived url would throw a 500 server error instead of 404.